Database ID:s finally changed from Integer to Long. Large restructuring of how RevisionData is accessed and handled. Row and SavedReference operations still need to be changed to the new format and old code access needs to be cleaned up to remove all direct access and inserting of values.

+++ metka/src/main/java/fi/uta/fsd/metka/data/entity/SequenceEntity.java
@@ -10,7 +10,7 @@ public class SequenceEntity {
     private String key;
 
     @Column(name = "SEQUENCE")
-    private Integer sequence;
+    private Long sequence;
 
     public String getKey() {
         return key;
@@ -20,11 +20,11 @@ public class SequenceEntity {
         this.key = key;
     }
 
-    public Integer getSequence() {
+    public Long getSequence() {
         return sequence;
     }
 
-    public void setSequence(Integer sequence) {
+    public void setSequence(Long sequence) {
         this.sequence = sequence;
     }
