+++ metka/src/main/java/fi/uta/fsd/metka/mvc/controller/SeriesController.java
@@ -55,7 +55,7 @@ public class SeriesController {
     * If no revision is found then return to search page with an error message.
     */
     @RequestMapping(value = "view/{id}", method = RequestMethod.GET)
-    public String view(Model model, @PathVariable Integer id, RedirectAttributes redirectAttributes) {
+    public String view(Model model, @PathVariable Long id, RedirectAttributes redirectAttributes) {
         Integer revision = seriesService.findSingleRevisionNo(id);
         if(model.asMap().containsKey("displayableErrors")) {
             redirectAttributes.addFlashAttribute("displayableErrors", model.asMap().get("displayableErrors"));
@@ -79,7 +79,7 @@ public class SeriesController {
     */
     @RequestMapping(value = "view/{id}/{revision}", method = RequestMethod.GET)
     public String viewRevision(Model model,
-                               @PathVariable Integer id, @PathVariable Integer revision,
+                               @PathVariable Long id, @PathVariable Integer revision,
                                RedirectAttributes redirectAttributes) {
         TransferObject single = null;
         Configuration config = null;
@@ -195,7 +195,7 @@ public class SeriesController {
     * actually required or is there already an open DRAFT revision).
     */
     @RequestMapping(value = "edit/{seriesno}", method = {RequestMethod.GET})
-    public String edit(@PathVariable Integer seriesno, RedirectAttributes redirectAttributes) {
+    public String edit(@PathVariable Long seriesno, RedirectAttributes redirectAttributes) {
         RevisionViewDataContainer revData = seriesService.editSeries(seriesno);
         if(revData == null || revData.getTransferObject() == null || revData.getConfiguration() == null) {
             // TODO: Notify user that no editable revision could be found or created
